Refactor: GPT Result Table 컴포넌트 분리
